Test Matching Data/Metadata Fields

test_fields_match() compares the attributes of each dataTable within the EML metadata to the columns in the corresponding .csv. If the columns have the same names and order, the test passes. If the columns differ, the test fails with an error.

Usage

test_fields_match(
  directory = here::here(),
  metadata = load_metadata(directory)
)

Arguments

directory

the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.

metadata

The metadata object returned by load_metadata. If parameter not provided, defaults to calling load_metadata in current project directory.

Value

Invisibly returns metadata.

Details

test_fields_match briefly checks that data files match, but you should really run test_file_name_match() before you run this test.

Examples

dir <- DPchecker_example("BICY_veg")
#> Data are provided for example use only. Do not assume that they are complete, accurate, or up to date.
test_fields_match(dir)
#> Called from: eval(expr, p)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#524: data_tbl <- EML::eml_get(metadata, "dataTable")
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#526: if ("attributeList" %in% names(data_tbl)) {
#>     data_tbl <- list(data_tbl)
#> }
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#531: metadata_attrs <- lapply(data_tbl, function(tbl) {
#>     arcticdatautils::eml_get_simple(tbl, "attributeName")
#> })
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#532: metadata_attrs$`@context` <- NULL
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#535: data_files <- list.files(path = directory, pattern = ".csv")
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#538: table_names <- NULL
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#539: for (i in 1:length(seq_along(data_files))) {
#>     tbl_nam <- data_tbl[[i]][["physical"]][["objectName"]]
#>     table_names <- append(table_names, tbl_nam)
#> }
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#540: tbl_nam <- data_tbl[[i]][["physical"]][["objectName"]]
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#541: table_names <- append(table_names, tbl_nam)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#540: tbl_nam <- data_tbl[[i]][["physical"]][["objectName"]]
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#541: table_names <- append(table_names, tbl_nam)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#540: tbl_nam <- data_tbl[[i]][["physical"]][["objectName"]]
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#541: table_names <- append(table_names, tbl_nam)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#544: names(metadata_attrs) <- table_names
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#547: data_colnames <- sapply(data_files, function(data_file) {
#>     names(readr::read_csv(file.path(directory, data_file), n_max = 1, 
#>         show_col_types = FALSE))
#> }, USE.NAMES = TRUE, simplify = FALSE)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#550: if (!(all(names(data_colnames) %in% names(metadata_attrs)) & 
#>     all(names(metadata_attrs) %in% names(data_colnames)))) {
#>     cli::cli_abort(c(x = "Mismatch in data file column names and the attributes listed in metadata."))
#> }
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#555: mismatches <- sapply(data_files, function(data_file) {
#>     meta_cols <- metadata_attrs[[data_file]]
#>     data_cols <- data_colnames[[data_file]]
#>     if (length(meta_cols) == length(data_cols) && all(meta_cols == 
#>         data_cols)) {
#>         return(NULL)
#>     }
#>     else if (all(meta_cols %in% data_cols) && all(data_cols %in% 
#>         meta_cols)) {
#>         return(c(` ` = paste0("--> {.file ", data_file, "}: Metadata column order does not match data column order")))
#>     }
#>     else {
#>         missing_from_meta <- data_cols[!(data_cols %in% meta_cols)]
#>         if (length(missing_from_meta) > 0) {
#>             missing_from_meta <- paste0("----Missing from metadata: ", 
#>                 paste0("{.field ", missing_from_meta, "}", collapse = ", "))
#>             names(missing_from_meta) <- " "
#>         }
#>         missing_from_data <- meta_cols[!(meta_cols %in% data_cols)]
#>         if (length(missing_from_data) > 0) {
#>             missing_from_data <- paste0("----Missing from data file: ", 
#>                 paste0("{.field ", missing_from_data, "}", collapse = ", "))
#>             names(missing_from_data) <- " "
#>         }
#>         msg <- c(` ` = paste0("--> {.file ", data_file, "}:"), 
#>             missing_from_meta, missing_from_data)
#>         return(msg)
#>     }
#> }, USE.NAMES = FALSE, simplify = FALSE)
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#579: mismatches <- purrr::discard(mismatches, is.null) %>% unlist()
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#583: if (!is.null(mismatches)) {
#>     cli::cli_abort(c(x = "Column mismatch between data and metadata.", 
#>         mismatches))
#> } else {
#>     cli::cli_inform(c(v = "All columns in data match all columns in metadata."))
#> }
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#586: cli::cli_inform(c(v = "All columns in data match all columns in metadata."))
#> ✔ All columns in data match all columns in metadata.
#> debug at C:/Users/rlbaker/Documents/RDev/DPchecker_devspace/R/tabular_data_congruence.R#589: return(invisible(metadata))